refactor(back): improve tests commands
