+++ back/.mocharc.unit.json
@@ -1,0 +1,12 @@
+{
+    "require": ["ts-node/register", "src/tests/unit/setup.ts"],
+    "node-option": [
+        "loader=ts-node/esm"
+    ],
+    "extension": ["ts"],
+    "spec": ["./src/tests/unit/**/*.test.*"],
+    "timeout": 2000,
+    "recursive": true,
+    "reporter": "spec",
+    "ui": "bdd"
+  }
