+++ back/src/tests/integration/setup.ts
@@ -1,0 +1,10 @@
+import { configDotenv } from "dotenv";
+import { execSync } from "node:child_process";
+
+configDotenv();
+
+process.env.NODE_ENV = "test";
+
+execSync("npm run db:prepare:test", {
+  stdio: "inherit",
+});
