+++ back/src/tests/unit/consumers/routes/deleteExpense.test.ts
@@ -1,14 +1,17 @@
 import http from "node:http";
 import request from "supertest";
-import sinon, { SinonStub } from "sinon";
+import sinon from "sinon";
 import { afterEach, beforeEach } from "mocha";
-import { authenticate, expect, mockedServer } from "./test-helpers.js";
-import * as deps from "../../../ioc.js";
-import AddOutflowCommand from "../../../core/commands/AddOutflowCommand.js";
-import { MonthNotFoundError } from "../../../core/errors/MonthErrors.js";
-import { Deps } from "../../../ioc.js";
+import {
+  authenticate,
+  expect,
+  mockedServer,
+} from "../../../integration/consumers/test-helpers.js";
+import * as deps from "../../../../ioc.js";
+import DeleteExpenseCommand from "../../../../core/commands/DeleteExpenseCommand.js";
+import { MonthNotFoundError } from "../../../../core/errors/MonthErrors.js";
 
-describe("Integration | Consumers | Routes | PUT /months/{monthId}/outflows", function () {
+describe("Integration | Consumers | Routes | DELETE /months/{monthId}/weekly/{weeklyId}/expenses/{expenseId}", function () {
   let server: http.Server;
 
   describe("When user is authenticated", function () {
@@ -18,73 +21,52 @@ describe("Integration | Consumers | Routes | PUT /months/{monthId}/outflows", fu
 
     it("should return a 200 on happy path", async function () {
       // given
-      const command: AddOutflowCommand = {
-        monthId: "a0c3861c-1383-4ef3-8c6b-167dcddd8f78",
-        label: "JOW",
-        amount: 10,
+      const command: DeleteExpenseCommand = {
+        monthId: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+        weeklyId: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+        expenseId: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
       };
       const updatedMonth = Symbol("updatedMonth");
       const dto = "dto";
-      const depsStub: Deps = {
+      const depsStub = {
         ...deps,
         monthDto: sinon.stub().withArgs(updatedMonth).returns(dto),
-        addOutflowDeserializer: sinon
-          .stub()
-          .withArgs(
-            { monthId: command.monthId },
-            { label: command.label, amount: command.amount }
-          )
-          .returns(command),
       };
-      depsStub.addOutflowUsecase.execute = sinon
+      depsStub.deleteExpenseUsecase.execute = sinon
         .stub()
         .withArgs(command)
         .resolves(updatedMonth);
 
       server = mockedServer({ isAuthenticated: true }, depsStub);
       const cookie = await authenticate(server);
-      const { monthId, ...body } = command;
+      const { monthId, weeklyId, expenseId } = command;
 
       // when
       const response = await request(server)
-        .post(`/months/${monthId}/outflows`)
-        .send(body)
+        .delete(`/months/${monthId}/weekly/${weeklyId}/expenses/${expenseId}`)
         .set("Cookie", cookie);
 
       // then
-      expect((depsStub.addOutflowDeserializer as SinonStub).args).to.deep.equal(
-        [
-          [
-            { monthId: command.monthId },
-            { label: command.label, amount: command.amount },
-          ],
-        ]
-      );
-      expect(depsStub.monthDto).has.been.calledOnceWith(updatedMonth);
-      expect(
-        (depsStub.addOutflowUsecase.execute as SinonStub).args
-      ).to.deep.equal([[command]]);
-      expect(response.statusCode).to.be.equal(201);
+      expect(response.statusCode).to.be.equal(200);
       expect(response.body.success).to.be.true;
       expect(response.body.data).to.be.equal(dto);
     });
 
     it("should return 400 error if body was malformed", async function () {
       // given
-      const command: AddOutflowCommand = {
+      const command: DeleteExpenseCommand = {
         monthId: "not-an-id",
-        label: "JOW",
-        amount: 10,
+        weeklyId: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+        expenseId: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
       };
 
       server = mockedServer({ isAuthenticated: true }, deps);
       const cookie = await authenticate(server);
-      const { monthId, ...body } = command;
+      const { monthId, weeklyId, expenseId } = command;
 
       // when
       const response = await request(server)
-        .post(`/months/${monthId}/outflows`)
-        .send(body)
+        .delete(`/months/${monthId}/weekly/${weeklyId}/expenses/${expenseId}`)
         .set("Cookie", cookie);
 
       // then
@@ -94,24 +76,23 @@ describe("Integration | Consumers | Routes | PUT /months/{monthId}/outflows", fu
 
     it("should return 404 error if a model is not found", async function () {
       // given
-      const command: AddOutflowCommand = {
-        monthId: "a0c3861c-1383-4ef3-8c6b-167dcddd8f78",
-        label: "JOW",
-        amount: 10,
+      const command: DeleteExpenseCommand = {
+        monthId: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+        weeklyId: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+        expenseId: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
       };
-      deps.addOutflowUsecase.execute = sinon
+      deps.deleteExpenseUsecase.execute = sinon
         .stub()
         .withArgs(command)
         .throwsException(new MonthNotFoundError());
 
       server = mockedServer({ isAuthenticated: true }, deps);
       const cookie = await authenticate(server);
-      const { monthId, ...body } = command;
+      const { monthId, weeklyId, expenseId } = command;
 
       // when
       const response = await request(server)
-        .post(`/months/${monthId}/outflows`)
-        .send(body)
+        .delete(`/months/${monthId}/weekly/${weeklyId}/expenses/${expenseId}`)
         .set("Cookie", cookie);
 
       // then
@@ -131,7 +112,9 @@ describe("Integration | Consumers | Routes | PUT /months/{monthId}/outflows", fu
 
     it("should return a 401 error", async function () {
       await request(server)
-        .put("/months/85467999-2751-4f04-8d96-7d7727fbff02/outflows")
+        .delete(
+          "/months/85467999-2751-4f04-8d96-7d7727fbff02/weekly/85467999-2751-4f04-8d96-7d7727fbff02/expenses/85467999-2751-4f04-8d96-7d7727fbff02"
+        )
         .expect(401);
     });
   });
