+++ back/src/tests/unit/consumers/routes/getYearlyOutflows.test.ts
@@ -2,8 +2,12 @@ import http from "node:http";
 import request from "supertest";
 import sinon, { SinonStub } from "sinon";
 import { afterEach, beforeEach } from "mocha";
-import { authenticate, expect, mockedServer } from "./test-helpers.js";
-import * as deps from "../../../ioc.js";
+import {
+  authenticate,
+  expect,
+  mockedServer,
+} from "../../../integration/consumers/test-helpers.js";
+import * as deps from "../../../../ioc.js";
 
 describe("Integration | Consumers | Routes | GET /yearly-outflows", function () {
   let server: http.Server;
