+++ back/src/tests/unit/consumers/routes/manageOutflowsChecking.test.ts
@@ -1,13 +1,17 @@
 import http from "node:http";
 import request from "supertest";
-import sinon, { SinonStub } from "sinon";
+import sinon from "sinon";
 import { afterEach, beforeEach } from "mocha";
-import { authenticate, expect, mockedServer } from "./test-helpers.js";
-import * as deps from "../../../ioc.js";
-import ArchiveMonthCommand from "../../../core/commands/ArchiveMonthCommand.js";
-import { MonthNotFoundError } from "../../../core/errors/MonthErrors.js";
+import {
+  authenticate,
+  expect,
+  mockedServer,
+} from "../../../integration/consumers/test-helpers.js";
+import * as deps from "../../../../ioc.js";
+import ManageOutflowsCheckingCommand from "../../../../core/commands/ManageOutflowsCheckingCommand.js";
+import { MonthNotFoundError } from "../../../../core/errors/MonthErrors.js";
 
-describe("Integration | Consumers | Routes | PUT /months/{monthId}/archive", function () {
+describe("Integration | Consumers | Routes | PUT /months/{monthId}/outflows/checking", function () {
   let server: http.Server;
 
   describe("When user is authenticated", function () {
@@ -17,39 +21,42 @@ describe("Integration | Consumers | Routes | PUT /months/{monthId}/archive", fun
 
     it("should return a 200 on happy path", async function () {
       // given
-      const command: ArchiveMonthCommand = {
+      const command: ManageOutflowsCheckingCommand = {
         monthId: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+        currentBalance: 2000,
+        outflows: [
+          {
+            id: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+            isChecked: false,
+          },
+          {
+            id: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+            isChecked: true,
+          },
+        ],
       };
       const updatedMonth = Symbol("updatedMonth");
       const dto = "dto";
       const depsStub = {
         ...deps,
         monthDto: sinon.stub().withArgs(updatedMonth).returns(dto),
-        archiveMonthDeserializer: sinon
-          .stub()
-          .withArgs(command)
-          .returns(command),
       };
-      depsStub.archiveMonthUsecase.execute = sinon
+      depsStub.manageOutflowsCheckingUsecase.execute = sinon
         .stub()
+        .withArgs(command)
         .resolves(updatedMonth);
 
       server = mockedServer({ isAuthenticated: true }, depsStub);
       const cookie = await authenticate(server);
-      const { monthId } = command;
+      const { monthId, ...body } = command;
 
       // when
       const response = await request(server)
-        .put(`/months/${monthId}/archive`)
+        .put(`/months/${monthId}/outflows/checking/`)
+        .send(body)
         .set("Cookie", cookie);
 
       // then
-      expect(
-        (depsStub.archiveMonthDeserializer as SinonStub).args
-      ).to.deep.equal([[command]]);
-      expect(
-        (depsStub.archiveMonthUsecase.execute as SinonStub).args
-      ).to.deep.equal([[command]]);
       expect(response.statusCode).to.be.equal(200);
       expect(response.body.success).to.be.true;
       expect(response.body.data).to.be.equal(dto);
@@ -57,17 +64,29 @@ describe("Integration | Consumers | Routes | PUT /months/{monthId}/archive", fun
 
     it("should return 400 error if body was malformed", async function () {
       // given
-      const command: ArchiveMonthCommand = {
+      const command: ManageOutflowsCheckingCommand = {
         monthId: "not-an-id",
+        currentBalance: 2000,
+        outflows: [
+          {
+            id: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+            isChecked: false,
+          },
+          {
+            id: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+            isChecked: true,
+          },
+        ],
       };
 
       server = mockedServer({ isAuthenticated: true }, deps);
       const cookie = await authenticate(server);
-      const { monthId } = command;
+      const { monthId, ...body } = command;
 
       // when
       const response = await request(server)
-        .put(`/months/${monthId}/archive`)
+        .put(`/months/${monthId}/outflows/checking/`)
+        .send(body)
         .set("Cookie", cookie);
 
       // then
@@ -77,21 +96,33 @@ describe("Integration | Consumers | Routes | PUT /months/{monthId}/archive", fun
 
     it("should return 404 error if a model is not found", async function () {
       // given
-      const command: ArchiveMonthCommand = {
+      const command: ManageOutflowsCheckingCommand = {
         monthId: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+        currentBalance: 2000,
+        outflows: [
+          {
+            id: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+            isChecked: false,
+          },
+          {
+            id: "60164d6e-3c17-43ed-bf3e-24e44e68d857",
+            isChecked: true,
+          },
+        ],
       };
-      deps.archiveMonthUsecase.execute = sinon
+      deps.manageOutflowsCheckingUsecase.execute = sinon
         .stub()
         .withArgs(command)
         .throwsException(new MonthNotFoundError());
 
       server = mockedServer({ isAuthenticated: true }, deps);
       const cookie = await authenticate(server);
-      const { monthId } = command;
+      const { monthId, ...body } = command;
 
       // when
       const response = await request(server)
-        .put(`/months/${monthId}/archive`)
+        .put(`/months/${monthId}/outflows/checking/`)
+        .send(body)
         .set("Cookie", cookie);
 
       // then
@@ -111,7 +142,7 @@ describe("Integration | Consumers | Routes | PUT /months/{monthId}/archive", fun
 
     it("should return a 401 error", async function () {
       await request(server)
-        .put("/months/85467999-2751-4f04-8d96-7d7727fbff02/archive")
+        .put("/months/85467999-2751-4f04-8d96-7d7727fbff02/outflows/checking")
         .expect(401);
     });
   });
