+++ back/src/tests/unit/setup.ts
@@ -1,0 +1,6 @@
+import { configDotenv } from "dotenv";
+
+configDotenv();
+
+process.env.NODE_ENV = "test";
+process.env.MOCHA_TEST = "unit";
